ENH: Finish "collapsed/summarized rows" feature. Automatically collapse rows with common site, buildname, stamp and group position values, but avoid collapsing results of filter queries. In collapsed rows, provide a link in the "Build Name" column to filtered results based on the Site Name and Build Name fields - allows user to drill down into the collapsed rows.
ENH: Add a "Create Hyperlink" button in the filter form. Once clicked, a hyperlink appears: users may right click it in their web browser to copy it to the clipboard or add it as a bookmark. Or just click it directly to verify that it works. Allows quickly reloading previously tediously entered filter parameters.