Fix problem retrieving daily updates with git repositories.
The "SELECT revision FROM dailyupdate" query was missing single quotes, and so could not successfully retrieve the prior revision number for generating the correct git 'whatchanged' command line.

Also, the "back-tick shell command" running technique used to execute git was not working on Windows with msys git because of double-quoting problems and spaces in the "Program Files" path. Fixed that, too.

Then: added 'revision' and 'priorrevision' and associated elements to the viewChanges.php page, so that it (a) displays that information to the end user, and (b) is the same as the viewUpdate.php page. They already share the display code in the xsl file; they should also share as much php code as possible, too. Also, add a line break after the priorrevision line, and only display priorrevision if we're also displaying revision, both for better readability.