Revamp Project.xml processing so that it never removes subprojects automatically. CDash has a one-mapping-per-project view of the subprojects at any given point in time. Before this change, if it received submissions from clients with divergent views, then we end up deleting subprojects according to client A's submission and then adding them back again according to client B's submission. This could (and did) lead to daily oscillations of the subproject dependency data.
It adds and removes relationships between subprojects still, but only for subproject relationships where both ends of the relationship have a SubProject element in the submitted Project.xml file.

This change avoids the problem of the disappearing subproject by never deleting subprojects automatically.

In the future, it may be desirable to re-instate the code for deleting subprojects, but only with additional information sent in the Project.xml file. Perhaps an autoRemove="true" attribute on the top level Project element.

It may also be desirable to further this work and never automatically remove relationships according to submissions. Time will tell on that one.

For now, if a subproject needs to be removed from a CDash project, the project admin should log into the CDash server and remove it manually.