Associate submissions with client job schedules when necessary.
This commit fixes a bug where a client submission may erroneously be considered
a non-client build.  This happened when CDash was busy processing non-client
builds when the client submitted its results to the dashboard.  This resulted
in an error due to the way we used to handle $clientscheduleid.  This value was
applied (or not) to the entire batch of submissions processed by
processsubmissions.php.  We fix this issue with a new database table:
client_jobschedule2submission.  This ties a submission to a scheduleid when
necessary.  This allows processsubmissions.php to handle a mix of client and
non-client submissions in a single batch.  This commit should cause the test
properclientprocessing to begin passing.