BUG: Remove 'exit' call and replace it with a trigger_error call.
The call to exit() would simply allow the submission processing loop to hang, not knowing that a non-recoverable error has occurred. Using trigger_error here instead allows the processing to halt immediately on the current file and move on to the next one. Done, done, on to the next one; done, I'm done and I'm on to the next...