ENH: Add config setting CDASH_LARGE_TEXT_LIMIT.
Default value is 0, meaning do not limit large text. (Same as before.)

If you set it, we will truncate large StdOutput and StdError values in BuildFailure instances before inserting them into the database.

The example XML file that I used to develop this code was 109M large, with 376 build warnings in it, each one averaging about 280k+ of "StdError"... We could not process it successfully until limiting the size to 261999 or less.

Given that, and the fact that we have some builds that report 2,000+ warnings, I would recommend setting this new config value to 32k AT MOST. In fact, it would be very good to limit it down to 8k or even 1k for most CDash servers.

Some people may complain about this limit, but it's better to have it and successfully process the file and have partial data than it is to not be able to process the file at all and have no data.