BUG: Eliminate excessive mem usage when viewing large numbers of build warnings or errors.
(This is the same change as revision 3004 in trunk.)

Do not call create_function on every single call to add_XML_value. It consumes 1640 bytes of memory that does not get reclaimed until php script end time...

Also, use a local variable inside the while loops in viewBuildError.php (lxml) to avoid frequent, looping, multi-megabyte string concatenations on every single .= in the file. Now, most of the .= ops are on the much smaller lxml variable except for one at the end of each while loop...

Now you can actually view a build error page with 2,000 plus build errors on it in about 10 seconds. Previously, we would simply run out of RAM and/or timeout because multi-megabyte strcats were taking too long.