ENH: new overview page
The project overview allows us to quickly see how the quality
of a project's key build group(s) have changed over the past
two weeks.

This commit also introduces manageOverview.php, which allows
project admins to determine which build groups should be
included in the overview, and in what order they should appear.

This commit also extends the concept of build groups.  These
can now be defined by substring matching on a build's name.
This option appears as a new tab on manageBuildGroup.

Finally, I updated manageBuildGroup to use the jquery-enabled
drag-and-drop sorting that I used as part of manageOverview.
This removes the need to refresh this page multiple times
when rearranging the positions of your build groups.