ENH: improve appearance of some new features
Add an option to make the X-axis of our line charts
(as seen on overview.php) to display dates.  Before,
they just said -1, -2, etc., for how many days in
the past this data point represented.

Also improve the display of core vs. non-core coverage
on index.php.  The rows now restripe nicely after collapse
or expand, and a folder icon displays this state for these
groups.