ENH: Generalize overview pages
This commit update the manage overview pages
to allow a full customization of the information
displayed.

The overview table has been changed accordingly.
The update script relies on a "DROP TABLE". This is a
sub-optimal but considering the feature is in its
infancy and used only by one project, this is a non issue.

It includes the minor UI tweaks. The "Save Layout"
button is more prominent. The bullet chart has also
been improved but the tooltips are still popping up in
the wrong spots.

From: Zack Galbreath <zack.galbreath@kitware.com>