BUG: fix load times on manageBuildGroup
This page was attempting to look up all the expected builds
for a project, but it was actually returning all builds (expected
or not), regardless of how long ago they were performed.  This
commit fixes this bug, resulting in improved behavior and performance.

This commit also introduces some improvements to the "define build
group by name" tab, which has been renamed "wildcard".  This pane
now displays which wildcard groups have already been defined.