BUG: fix broken postgres installation
This was caused by a trailing comma in pgsql/cdash.sql.

I also modified the postgres testing code to check for
a default database named 'postgres' when attempting to
recreate the testing database.  Previously the only
name that was checked was 'host'.

This commit also fixes a typo 'sucess' that appeared
during the installation process.