Remove build name from default nightly sort list
Before this commit, rows for expected builds that did not yet submit
were scattered throughout their group.  This is due to the fact that
by default, we are sorting by build name.  I'm not 100% sure, but
my guess is that this change in behavior from previous versions of
CDash is due to a jquery version upgrade.