Fix parsing of JSCover JSON Files
The JSCover json files may have information more than 1 JS file in
the content.  Change the parser to  generate a coverage object for
each entry in the json files.

Pass two arrays to the Parse function to keep track of files found
in multiple JSON files.

Change "is_numeric" to "isset" for finding non-runnable lines.  Check
in the passed arrays for an existing coverage object before generating
a new one.

Change the coveragefilelog "insert" function to use += instead of =
for multiple entries of the same file.