bug fix for JSCover parser
This commit fixes a bug in the JSCover parser that prevented
it from accumulating coverage information properly.
If multiple JSON files contained coverage data for a single
source file, the final JSON file to be parsed would overwrite
any coverage data contained in the previous ones.

This commit also encapsulates the recursive DeleteDirectory
function so as to reduce code duplication.