fix failing tests
The introduction of CDASH_MAX_QUERY_RETRIES caused a handful of tests
to fail.  Apparently we expect or tolerate some queries to fail.
As far as I could tell, there is no way to configure simpletest to
wait longer for a response, so for now we redefine the default
value of CDASH_MAX_QUERY_RETRIES to 1.