enable reverse sorting per field
Make it possible to reverse sorting on a per-field basis.
This commit encapsulates the multi-sort functionality so
this code can be shared by different controllers.

This commit also makes descending order the default for
sorting.  The idea here is that users will most likely want
to see builds with high numbers of errors / failing tests / etc.
at the top of the list.