Use directives to improve rendering of build rows
The repeated use of ng-if="numchildren...." was causing index.php
to render slowly when lots of builds were present.  To help alleviate
this problem, I've separated out the parent vs "normal" build case into
two separate templates.

These are called using an Angular directive rather than ng-include because
this seems a bit faster and causes less of a "flicker" effect when the
page is first loaded.