Make sure update_handler finds parent builds
When searching for an existing build within update_handler
defer to the parent if a child build is initially found.
The assumption here is that a parent build will update once
before all of its children are built.

This fixes a problem revealed in testing where child build
start times were sometimes set to too early of a value by
update_handler.