Assign coverage to correct version of the file
The previous solution here was too broad.  It could result in
coverage being associated with an older version of the file.

If we're not sure which version of the file is being covered
it is better to insert an empty row and let the coveragefile
object figure it out later.