Use one-time binding to speed up render of index.php
Use AngularJS's one-time binding feature (indicated by '::')
to reduce the number of watchers we generate.  This is helpful
when viewing lots of builds on index.php.

One-time binding only works if the expression does not evaluate
to undefined.  Thus we update the API endpoint to make sure that
all the expressions used in the view are defined, even if they
just evaluate to false.