Speedup coverage aggregation for GcovTar_handler
To assign aggregate coverage to the correct build we need to know
the beginning and end of the current testing day, as well as
the siteid of 'CDash Server'.

These values are not specific to the covered source file.
So rather than looking them up in CoverageFileLog::UpdateAggregate(),
we now find them in GcovTar_Handler instead.  This means that we are
only finding this data once per tarball instead of finding it for each
.gcov file.