Standardize naming scheme
In some places we were using 'nice_name' to refer to a human readable
string and 'name' to refer to the sanitized version.  In other places
we were using 'name' and 'name_clean', respectively.  This commit
standardizes on the latter convention.  A benefit of this change is
that overview now says "configure errors" rather than "configure_errors".

This commit also horizontally centers the text fields in the dynamic &
static analysis sections.