Carefully handle low precision time in note_handler
The DateTime field in Note.xml only has precision in minutes (not seconds).

If the build already exists, don't update its StartTime or EndTime
to this value.

If the build doesn't already exist we have no choice but to use
this time value.  Add 59 seconds to it to initialize the build's
start time.  This way any subsequent XML file with a more precise
start time will take precedence.