Zero out coverage summary diff when appropriate
Our GcovTar handler updates the coverage difference between
this build and the previous one on a file-by-file basis.

Because we refused to insert an entry for no difference, we could
end up with a situation where CDash would report a change in coverage
when none actually existed.  Instead, we now allow such "no difference"
entries to be created to override an existing difference that is no
longer valid.