Rename 'Build Time' filter to 'Build Start Time'
The term 'Build Time' was ambiguous.  Did it refer to when the
build occurred, or how long it took?

We opt for 'Build Start Time' rather than the less verbose 'Start Time'
because this filter is available on both index.php (which displays builds)
and queryTests.php (which displays tests).