Restructure configure table
The motivation behind this work is to allow multiple builds to
share a single configure step.  This will be used when splitting
a single build across multiple SubProjects.

I'm also taking this opportunity to reduce the size of our database
by moving the fields that are most likely to change (starttime and endtime)
to the new build2configure table.  This allows more builds to share
a configure row, reducing the need for duplicates in the expensive
'command' and 'log' columns.