Query GitHub API for version-only updates
When CTEST_UPDATE_VERSION_ONLY is set, CTest records the current
version of the repository without performing an update.  This is
useful when some other tool is responsible for pulling in changes.
But in this case CDash doesn't know what files changed, so it
can't assign credit to users for fixing warnings, errors, and tests.

This commit teaches CDash to query the GitHub API to gain this information.
It also fixes some bugs in the way user statistics are calculated.