Change how we count configure errors on viewSubProjects.php
We now count distinct configures that have errors as opposed to
counting builds with configure errors.  This better handles the
case where multiple (subproject) builds share a single configure step.
Under the previous approach, if a single shared configure had an error
it would be reported as N errors (one for each subproject).