Use get_seconds_from_interval for SubProject filters
The build_survives_filter function is used to reevaluate our
filters (build duration, test failures, etc) when we are including
or excluding SubProject(s).

Within this function we need to convert interval strings ('10m 45s')
to a number of seconds (645).  This number of seconds is then used
to check whether or not a build should be filtered out.

As part of this commit we also strengthen the ExcludeSubProjects test
to verify that this bug stays fixed.