Improve how we delete tests
For high-traffic CDash installations the build2test table can
contain 100 million rows or more.  Because of this potentially
huge size, attempting to join this table to itself can take a
very long time.

To reduce the amount of time that remove_build takes to complete,
we break this query up into a few easier to process steps.