Fix deficiency in SubProjects filter
Before this change, the SubProjects filter could return inconsistent
results when used to expand the set of builds returned (rather than
narrowing them down).  This occurred because the filter-modified SQL
query would prune results that should be re-added later by the
SubProjects filter.

To fix this problem, we now check for the presence of any SubProjects
filter before modifying the underlying SQL query.  If one is found,
then we do not add SQL clauses for other filters that maybe impacted.
Instead, we rely on the build_survives_filter function to correctly
prune results after the SubProjects filter has been applied.