Cancel test summary AJAX query when user navigates away
This fixes the issue of testDetails.php sometimes taking a long time to load.

viewTest.php's Summary column shows how tests performed for other builds
in the same group.  This can take a while to look up, so we perform this
step as an AJAX call after the initial page has loaded.

PHP's MySQL module only allows you to perform one query at a time though.
So if the user clicks on a link, this new page will have to wait for the
AJAX call to complete.

We fix this behavior by terminating the AJAX call whenever a user clicks
one of the links on viewTest.php.