New function: pdo_execute
I noticed that we were not logging SQL errors from prepared statements.
This was happening because the PDO object does not have access to errors
that occur in prepared statements.

pdo_execute() is a wrapper around PDOStatement::execute() that logs any
errors that occur during the execution of the prepared statement.