Missing Tests Now Display in UI and Email and Timeouts Now Display in Email
This addresses the issue of a test that has timed-out just shows up as “Failed” in the email notification. Now the email clearly indicates that the test timed-out.

This also adds support for missing tests. If tests go missing relative to the previous build, those tests will be clearly indicated in the emails and are also displayed in the UI.