Support for Disabled tests
A new test property in CMake will allow developers to mark a test as disabled.
Disabled tests are another type of Not Run test.  Internally, this distinction
is stored in the Details field of the test.

viewTest.php has been updated to show the details column for Not Run tests.
Disabled tests are shown in a different color on this page.

CDash's summary email has been updated to not include disabled tests in its list
of Not Run tests.