Use a stronger algorithm to hash passwords
We now use PHP's password_hash and password_verify functions to hash passwords
before storing them in the database.  This change requires us to increase the
length of the password fields in our database to 255 characters.
Old passwords hashed with md5 will be re-hashed when the user next logs in.

As part of this change, we have updated the User model to use PDO to interact
with the database.  We now use of the User model whenever necessary as opposed
to directly querying the database.