Build2Configure upgrade: compute crc32 in SQL
Use MySQL's CRC32 function when restructuring the configure table.
This is much faster than updating each configure row individually.

For Postgres, this function is not available, so we still compute the crc32
using PHP instead.  We now break up our "SELECT FROM configure" query into
chunks so as not to exhaust the available memory.