Don't check old groupid when moving a build
Somehow we ended up in a state where old build2grouprule rows were not getting
marked as finished when moving a build to a new group.  The result was two
active rules matching the build, with the oldest one winning out due to database
ordering.  This meant that builds would not move even after an admin told them
to.

We make this behavior more robust by not checking the old groupid when moving
a build to a new group.