Display link to parent notes on index.php
For our new method of submitting SubProject builds all at once, any notes
submitted get assigned to the parent build, not the individual children
SubProject builds.  In this case, we now display a link to such Notes at
the top of the page, alongside other bits of information that are common
among the children.