Prevent Project handler from overwriting user permissions
We noticed a bug where users were being stripped of their administrative
access to a project.  This was tracked down to the Project.xml parser.
If a user was listed as a contact for a SubProject (via an <Email> element)
then their subscription to the project would be replaced with a default
(non-administrative) role.

We now check for a previously existing subscription before saving one
with default permissions.