Don't show Time columns for all-at-once SubProject builds
When "Advanced View" is enabled, we typically see "Time" columns
for the Update, Configure, Build, and Test sections on index.php.
But for all-at-once SubProject builds, these columns were displaying
identical information repeated for each build.

This commit causes these Time columns to no longer be displayed
when viewing the results of an all-at-once SubProject build with
Advanced View enabled.  A subsequent commit will move this information
to the section at the top of the page that shows all the data that
is consistent across the SubProject builds.