Move Time columns to the shared header for all-at-once builds
In a previous commit, we removed the Time columns from Advanced View for
all-at-once builds.  We did this because this information does not vary
from build to build.  As a follow-up, we relocate this information to the
section at the top of the page containing data that is shared among the
children builds.