continue instead of return in sendemail()
Now that we suppose all-at-once SubProject builds, an XML handler could
generate emails for more than one build.

We should `continue` (rather than `return`) when we encounter a build that
has no errors.  When returning, it was possible to not send an email for a
build with error(s) if it appeared after a clean build in the list returned
by getActionableBuilds().