Add test case for SubProject build diffs
We noticed a bug where the +/- subscripts were not being displayed
when they should have been for some SubProject builds.  This bug was fixed
in the previous commit (c1ea4c7).  Now we add a test case to make sure
these subscripts are shown when they should be.

To do so, we modify our SubProjectNextPrevious test data so that the
Didasko build displays differences between the second and third day
(+1 warning, -1 notrun test, +1 passed test)  We also modify the start times
of our SubProject builds so that Mesquite is built before Didasko on the
third day.  This allows us to reproduce the original bug and make sure that
it stays fixed.