Show graph by default
When clicking on the 'Time Status' link, the test timing chart will now be shown
by default when testDetails.php is loaded.

Other links similarly improved by this commit include Previous/Next/Current
(at the top of the page), and the links you follow when clicking on individual
dots in these graphs.