Insert coverages one at a time to improve race condition
This decreases the likelihood of a race condition when doing
parallel processing of coverage. It doesn't get rid of the race
condition altogether (we would need some kind of locking or improved
upsert to do this), but it does significantly decrease the window of
time in which it can happen. We will need to figure out a better
long term solution.