fix: fixes 'transaction is aborted' on OSX PDOException: SQLSTATE: 25PO2 Driver code: 7 Message: ERROR:  current transaction is aborted, commands ignored until end of transaction block
The order of the prepared statements here seems to matter. When the statement that was moved into the else block was prepared just prior to prepared statement inside the if Exists() block the above error is thrown onOSX. Moving the first prepared statement so that one is prepared vs. the other seems to fix the problem.