session_set_cookie_params: only set lifetime
Change how session_set_cookie_params() is called so that we only pass in
first parameter (lifetime).  This is how the function was called in CDash
before the recent addition of the Session class.  We were concerned that
restricting the cookie to a certain domain may cause login to fail on some
CDash installations.