Only expire build rules from the same project
This commit fixes a bug where builds could unexpectedly get moved out of the
group that they were previously assigned to.  This would happen when you had
two builds with the same name and from the same site submitting to two
different projects.  When one of the project's build rules was updated,
it would affect the other project.  We now check what project a build rule
is for before marking it as finished.