refactor: prevent superflous fwd slash in url
If the server/site is not setting a value for CDASH_DIR_NAME a fwd
slash is appended to the server url. This causes URLs in testing
to look like, for instance, http://cdash.site//viewTests.php; a
url that causes problems when trying to use it with php functions
like parse_url.