Move directory trimming out of get_github_diff_url
get_github_diff_url previously removed the source directory from the
beginning of a path if it was found.  This caused problems on
viewUpdate.php, where we already separately perform a similar directory
pruning step.

To fix this problem, we make the directory trimming an explicit, separate
step and remove it from get_github_diff_url().