Make it possible to remove build filters
Prior to this commit, if you added a warning or error filter for your project
it was not possible to replace it with an empty entry on the Project Settings
page.

This commit also contains some reorganization of the BuildErrorFilter code.
It now accepts a Project model as a constructor argument as opposed to a
projectid, and it automatically loads its filters at construction time.
It also now exclusively uses PDO to interact with the database.