Fix bug where sub-filters wouldn't be displayed sometimes
This fixes a bug in the interaction between the SubProjects filter
and filter sub-blocks.

When filtering on SubProjects we remove some SQL WHERE clauses
that would otherwise appear in our queries and instead apply this
filtering in server-side PHP code instead.

Even though these filters do not correspond to any SQL clauses they
should still be displayed in the UI.