Soft delete dynamic buildgroup rules
Mark them as finished rather than deleting from the database.

Now that we have a timeline on viewBuildGroup.php, we would like to be able
to see how a dynamic buildgroup has performed and changed over time.
This change will allow you to see how well a buildgroup was performing in the
past based on the rules that were defined for it back then.