Show proc time for non-children builds on index.php
Prior to this commit, the "proc time" column was only shown on index.php
when the parentid parameter was specified.

This commit also changes the behavior of the SubProject filter on this page.
Previously it would modify the value of "Test Time". Now it leaves "Test Time"
alone but changes the value of "Proc Time".

This is in line with our redefinition of these fields.
"Test Time" is wall clock time for running all the tests.
"Proc Time" is the sum of each test's execution time multiplied by how many
processors it used.