Append 'UTC' to input of Project::GetTestingDay()
The input to this function is always a timestamp from the database,
which is stored in UTC format. So we need to explicitly tell PHP
what timezone this date string is from so that it can convert it
to the correct UNIX timestamp value. This is a bug fix for servers
that are configured to locally use a non-UTC timezone.