Fix parsing of XML files with lots of subprojects
Prior to this commit, CDash's XML parser would sniff through the first
8 kilobytes of an XML file to figure out what type of submission it was.
This assumption breaks down when we have lots of subprojects included in the
XML file, as these appear before the type-specific XML element
(<Build>, <Configure>, <Test>, etc.) that CDash is looking for.

The fix is to abandon the assumption that this identifying element will
appear within the first 8 kilobytes. Instead, we now stream through the XML file
from the beginning, 8KB at a time, trying to figure out what type of file it is.