Improve support for redundant tests
This commit fixes a bug on the 'test details' page.

It would occur when a build had two separate tests with the same name.
In that case, 'test details' would show the output from the 1st test, even
if you were attempting to look at the results from the second one.