Expand flags for composer process
Expand the flags used for the composer update/install process by adding
arguments to the update.  It should add a "no-dev" flag to the default
list and ensure that the versions are instructed to prefer the lowest
version of dependencies.

Add in a "-u" flag which isn't used in an automated fashion since it may
introduce breaking dependency issues.

Fixes: #1294