Add dependency audit command
Add a command which prints the reports of composer audit and npm audit to
the console.

Use Laravel's scheduler to run the task and pipe the output to a known
file where the maintenance page will read it in for display.

The task should be set to run every 6 hours to shorten the length of
time that a file might not exist.

Add buttons for clearing the file and the skeleton for running the
task on demand.