Update host name inside the testing Docker container
By default we refer to our testing CDash container as
http://localhost:8080 because this allows users to interact
with CDash using a web browser.

This approach causes our protractor tests to fail though,
as the Chrome driver runs in a separate container. For this reason,
we updates CDash's URL to be http://cdash:8080 while tests are running,
and then change it back to localhost:8080 afterwards.