Improve config/docs for production postgres
For now, this seems to be the best we can do.

Once docker compose v2.19 is released we should be able to
update the worker service by "extending" the cdash service,
removing the need for manually editing the production config YAML.

In particular, we need the new "!reset" feature to prevent both the web
service and the worker from attempting to bind port 443.
https://github.com/compose-spec/compose-go/pull/380