Refactor `buildProperties.php` API endpoint (#1524)
This PR is part of an ongoing effort to migrate all of our routes to
Laravel's routing/controller system.

In addition to moving the core logic of the `buildProperties.php` API
endpoint to a new `BuildPropertiesController`, I performed a light
refactor and fixed an issue which allowed unauthorized users to view
build output. Future work should be done to further clean up the code.