Refactor `addNote.php` API endpoint (#1528)
This PR is part of our ongoing effort to migrate all of the API
endpoints to Laravel's routing/controller system.

As an aside, it is *incredibly* confusing that we have both "user notes"
and "build notes", both of which are referred to as "notes", and tested
in the same test file. To differentiate between the two types of notes,
I have split the testing code into two separate files which each test
different kinds of notes.