Move `dailyupdatescurl.php` AJAX endpoint logic to controller (#1538)
By migrating the last `/ajax/*` route to a Laravel controller, this PR
marks the end of our ongoing effort to modernize all of the `/ajax/*`
routes. This route should probably be eliminated in the future, but I
have moved it to a controller temporarily so the AJAX-route-handling
logic can be eliminated entirely.