Display test details command line in monospace font (#1548)
The test details command line is currently displayed in regular text,
which can be difficult to read when it consists of a long string of
symbols. This PR changes the UI to match other parts of the codebase in
which text of this nature is displayed as prepared text in a monospace
font.