Refactor `Coverage` model (#1539)
This PR serves as a trial run of my new approach regarding the
modernization of large models for which is is infeasible to replace the
model entirely. Rather than taking the reflection approach used for the
`User` classes, I have instead taken an "adapter" approach, such that
the public-facing interface of the original `Coverage` class has not
changed, but all of the database interactions happen through the
public-facing interface of the new eloquent-based `Coverage` class.