Standardize page titles (#1555)
#1351 created a standardized page title system, but failed to update a
handful of the page titles which are set via an API response instead of
a Blade template. The ultimate goal of course is to set the page title
in some sort of centralized location, but that work will occur at a
later point. In the meantime, this PR fixes all of the issues for the
moment so the 3.2 release at least ships with something reasonable.