Fix `editSite.php` error when no sites claimed (#1562)
Fixes #1560.

It's worth noting that the edit site page as a whole is in desperate
need of an overhaul and I plan to refactor it sometime soon. This PR
only attempts to fix the error noted in #1560, and does not make an
attempt to fix all outstanding issues with this page.