Reformat build history table (#1563)
The build history table on `/build/<buildid>` currently lacks padding
which makes it look unpolished. This PR applies the same styling used
for the "current build" table to the build history table. The header of
the table and the "History" banner are now the same color though, which
could probably be cleaned up in the future. In fact, a light UI refresh
for the entire page should probably happen eventually since there is a
lot of wasted space and confusing UI elements.

Before:

![image](https://github.com/Kitware/CDash/assets/16820599/116a01aa-ca66-49dd-a092-7281d061b86c)

After:

![image](https://github.com/Kitware/CDash/assets/16820599/f476a84b-1738-4aa1-ae3d-81e5b56f032f)