Refactor `overview.php` UI (#1564)
`overview.php` is currently unusable due to various UI components
overlapping with one another. This PR reformats the page so all of the
available information is visible. There are still a number of other
issues with generating the data for this page on the server side, but
they don't directly relate to the UI issues fixed here and will be
addressed in separate PRs.

Before:

![image](https://github.com/Kitware/CDash/assets/16820599/38fca420-3b05-47de-92f7-165d5d8a6534)

After:

![image](https://github.com/Kitware/CDash/assets/16820599/dbae5c28-45a4-4ab0-be15-9581026cbb39)