Remove LOC displays from coverage graph (#1576)
To eliminate an error where the Lines Of Code data attempted to access a
non-existant yaxis object, remove the plots for LOC from the coverage
over time display.

With an update to the design, they may be reintroduced by switching to
`yaxes` in the plot's `options` object and then reintroducing the data.

Previous State:


![image](https://github.com/Kitware/CDash/assets/542106/5b6a4c45-e6a0-4e0e-a4df-0bea3f4de2ca)

Current State:

![image](https://github.com/Kitware/CDash/assets/542106/d8df0b3c-2051-45e4-b47e-9a1b7ada8525)