Index `(testid,buildid)` pairs in `build2test` table (#1577)
The `build2test` table is a pivot table which is used to relate builds
and tests. The most common use case involves joining the `build2test`
table with the `build` and `test` tables via its `testid` and `buildid`
columns. Each individual column in the `build2test` table is indexed,
but using any relation between testids and buildids requires the
database to load the entire `build2test` row. This is the primary cause
of the widespread slowness on the test summary page, amongst other
things.

I have added indices on `(buildid,testid)` and `(testid,buildid)`.
During a trial run on a large development database, I observed a
significant speedup on `testSummary.php`, going from 60+ seconds to less
than one for an average page load. Due to the size of the `build2test`
table, CDash administrators should be aware that creating these indices
may take a significant amount of time. On a large development database,
it took more than 20 minutes.