Add index to improve `/projects` performance (#1579)
The `/projects` page effectively serves as the homepage for CDash. Prior
to this change, the API for `/projects` took ~1s to load on a large
development instance. After adding this index, the page load time
dropped to an almost negligible 0.1 seconds. Rather than scanning
through an entire section of an index to find the maximum submit time,
the database can now find the maximum submit time for a given project
directly.