Convert echoed error messages to thrown exceptions in models (#1580)
As we gradually move the codebase to Laravel, we want to control all
output through Laravel alone. A common pattern throughout the CDash
codebase is the usage of the echo command to output error messages,
which unintentionally conflicts with other output such as JSON output in
an API response. Furthermore, these error cases are not logged, so CDash
administrators are not automatically alerted if users are repeatedly
experiencing an error.

In this PR, I have switched all of the echoed error messages to thrown
exceptions via Laravel's `abort()` helper. These exceptions will show up
in the logs (if configured), and the exceptions will be forced through
the same centralized error handling system for greater consistency
across pages.